Track bullets used, game state, and timing updates
Added a new integer member `bulletsUsed` to the `Player` class in `player.h` to track the number of bullets used by the player. Initialized `bulletsUsed` to 0 in the `Player` constructor and incremented it when a bullet is fired.

In `OpenGL3DTemplate.cpp`, included additional headers `<string>` and `<ctime>`, and added `using namespace std` and `using namespace std::chrono`. Introduced new boolean variables `isGameLose` and `isGameWin` to track game state. Added `steady_clock::time_point startTime` and a constant `GAME_TIME` to manage game timing.

Implemented a `reset` method in the `Cam` class to reset the camera position. Added a `renderBitmapString` function to render text on the screen and a `getTime` function to calculate elapsed time since the game started. Modified the `Display` function to show win/lose messages and objectives, and to reset the camera when the game ends. Displayed the remaining game time and objectives on the screen in the `Display` function. Checked for game win condition (using 3 bullets) and game lose condition (time up) in the `Timer` function. Initialized `startTime` in the `game_init` function. Changed the background color in the `main` function to a light blue color.

Reduced the `colorChangeInterval` from 2000 ms to 1000 ms in `renderWall.cpp`.